Centralise Cluster Explorer API implementation (#872)
